Revise morseToTwiML to make more use of the Twilio Go Helper Library
This is a small revision to have morseToTwiML return a slice of
twiml.Elements, instead of a custom TwiML structs. It also refactors
handleVoiceRequest to account for this change, and removes the TwiML and
Play structs, as they're no longer required.

The reason for this is that Twilio's Go Helper Library already does what
the code did. So, it makes sense to use it — especially when others are
maintaining the code, catching bugs, etc. The resulting TwiML is
DTD-compliant, and shorter, as it uses short, not long tags.