Streamline definition of the ServeMux
This change is comprised of two small changes. It declares a custom
ServeMux, instead of using the default, as that has security
implications as it can be overridden by any module or other element of
code.

It also makes use of changes in Go 1.22, where you can specify the
allowed HTTP method instead of using an if/else block to do so. See
https://matthewsetter.com/restrict-allowed-route-methods-go-122/ and
https://go.dev/blog/routing-enhancements#enhancements for further
details.